improve hce

+++ src/policy/abprune_hce.hpp
@@ -5,7 +5,7 @@
  * @brief alpha-beta prune + minimax + handcrafted evaluation
  */
 class ABPruneHCE {
-    static int _evaluate(State *state, int depth, int alpha, int beta, bool isMax);
+    static int _evaluate(State *state, int depth, float alpha, float beta, bool isMax);
 
 public:
     static Move get_move(State *state, int depth);
